+++ src/main/java/com/example/health_management/domain/repositories/HealthProviderRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HealthProviderRepository extends JpaRepository<HealthProvider, Integer> {
+public interface HealthProviderRepository extends JpaRepository<HealthProvider, Long> {
 }
